+++ src/main/java/pl/kupniewski/application/FoodStepSimulator.java
@@ -6,7 +6,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 import pl.kupniewski.application.usecases.deliver.DeliverOrderToCustomer;
 import pl.kupniewski.application.usecases.order.CreateOrderByCustomer;
-import pl.kupniewski.application.usecases.order.PreparedOrder;
+import pl.kupniewski.application.usecases.order.model.PreparedOrder;
 import pl.kupniewski.application.usecases.order.model.Order;
 import pl.kupniewski.application.usecases.restaurant.SimulatedRestaurant;
