+++ src/main/java/pl/kupniewski/application/usecases/deliver/DeliverOrderToCustomer.java
@@ -1,8 +1,7 @@
 package pl.kupniewski.application.usecases.deliver;
 
 import org.springframework.stereotype.Component;
-import pl.kupniewski.application.usecases.order.PreparedOrder;
-import pl.kupniewski.application.usecases.order.model.Order;
+import pl.kupniewski.application.usecases.order.model.PreparedOrder;
 
 @Component
 public class DeliverOrderToCustomer {
