+++ src/main/java/pl/kupniewski/application/usecases/restaurant/Restaurant.java
@@ -1,7 +1,6 @@
 package pl.kupniewski.application.usecases.restaurant;
 
-import pl.kupniewski.application.usecases.deliver.Deliverer;
-import pl.kupniewski.application.usecases.order.PreparedOrder;
+import pl.kupniewski.application.usecases.order.model.PreparedOrder;
 import pl.kupniewski.application.usecases.order.model.Order;
 
 public interface Restaurant {
